--- app/src/main/java/com/itmo/wineup/features/catalog/VineModel.kt
@@ -1,4 +1,0 @@
-package com.itmo.wineup.features.catalog
-
-data class VineModel(val name: String) {
-}
