+++ app/src/main/java/com/itmo/wineup/features/catalog/models/VineModel.kt
@@ -1,0 +1,4 @@
+package com.itmo.wineup.features.catalog.models
+
+data class VineModel(val name: String) {
+}
