+++ app/src/main/java/com/itmo/wineup/features/favorites/presentation/FavoritesFragment.kt
@@ -1,19 +1,18 @@
-package com.itmo.wineup.features.main
+package com.itmo.wineup.features.favorites.presentation
 
 import android.os.Bundle
-import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
 import com.itmo.wineup.R
-import com.itmo.wineup.features.favorites.FavoritesFragment
-import com.itmo.wineup.features.favorites.FavoritesViewModel
 
-class MainFragment : Fragment() {
+class FavoritesFragment : Fragment() {
 
     companion object {
-        fun newInstance() = MainFragment()
+        fun newInstance() =
+            FavoritesFragment()
     }
 
     override fun onCreateView(
@@ -21,14 +20,14 @@ class MainFragment : Fragment() {
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        return inflater.inflate(R.layout.fragment_main, container, false)
+        return inflater.inflate(R.layout.fragment_favorites, container, false)
     }
 
-    private lateinit var viewModel: MainViewModel
+    private lateinit var viewModel: FavoritesViewModel
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-        viewModel = ViewModelProvider(requireActivity()).get(MainViewModel::class.java)
+        viewModel = ViewModelProvider(requireActivity()).get(FavoritesViewModel::class.java)
     }
 
 }
