+++ app/src/main/java/com/itmo/wineup/features/main/presentation/MainFragment.kt
@@ -1,4 +1,4 @@
-package com.itmo.wineup.features.profile
+package com.itmo.wineup.features.main.presentation
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -7,13 +7,12 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.lifecycle.ViewModelProvider
 import com.itmo.wineup.R
-import com.itmo.wineup.features.main.MainFragment
-import com.itmo.wineup.features.main.MainViewModel
 
-class ProfileFragment : Fragment() {
+class MainFragment : Fragment() {
 
     companion object {
-        fun newInstance() = ProfileFragment()
+        fun newInstance() =
+            MainFragment()
     }
 
     override fun onCreateView(
@@ -21,14 +20,14 @@ class ProfileFragment : Fragment() {
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        return inflater.inflate(R.layout.fragment_profile, container, false)
+        return inflater.inflate(R.layout.fragment_main, container, false)
     }
 
-    private lateinit var viewModel: ProfileViewModel
+    private lateinit var viewModel: MainViewModel
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-        viewModel = ViewModelProvider(requireActivity()).get(ProfileViewModel::class.java)
+        viewModel = ViewModelProvider(requireActivity()).get(MainViewModel::class.java)
     }
 
 }
