+++ app/src/main/java/com/itmo/wineup/features/main/presentation/MainViewModel.kt
@@ -1,4 +1,4 @@
-package com.itmo.wineup.features.main
+package com.itmo.wineup.features.main.presentation
 
 import androidx.lifecycle.ViewModel
