+++ app/src/main/res/layout/fragment_catalog.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".features.catalog.CatalogFragment">
+    tools:context=".features.catalog.presentation.CatalogFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/vineListRecycler"
