+++ app/src/main/res/layout/fragment_favorites.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".features.favorites.FavoritesFragment">
+    tools:context=".features.favorites.presentation.FavoritesFragment">
 
     <TextView
         android:layout_width="match_parent"
