+++ app/src/main/res/layout/fragment_main.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".features.main.MainFragment">
+    tools:context=".features.main.presentation.MainFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
