+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -6,22 +6,22 @@
     app:startDestination="@+id/navigation_catalog">
     <fragment
         android:id="@+id/navigation_main"
-        android:name="com.itmo.wineup.features.main.MainFragment"
+        android:name="com.itmo.wineup.features.main.presentation.MainFragment"
         android:label="@string/main_title"
         tools:layout="@layout/fragment_main" />
     <fragment
         android:id="@+id/navigation_catalog"
-        android:name="com.itmo.wineup.features.catalog.CatalogFragment"
+        android:name="com.itmo.wineup.features.catalog.presentation.CatalogFragment"
         android:label="@string/catalog_title"
         tools:layout="@layout/fragment_catalog" />
     <fragment
         android:id="@+id/navigation_favorites"
-        android:name="com.itmo.wineup.features.favorites.FavoritesFragment"
+        android:name="com.itmo.wineup.features.favorites.presentation.FavoritesFragment"
         android:label="@string/favorites_title"
         tools:layout="@layout/fragment_favorites" />
     <fragment
         android:id="@+id/navigation_profile"
-        android:name="com.itmo.wineup.features.profile.ProfileFragment"
+        android:name="com.itmo.wineup.features.profile.presentation.ProfileFragment"
         android:label="@string/profile_title"
         tools:layout="@layout/fragment_profile" />
